+++ pyannote/audio/pipeline/speaker_diarization.py
@@ -26,326 +26,167 @@
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
 
-import numpy as np
-import sklearn.cluster
 from pathlib import Path
-from scipy.spatial.distance import squareform
-from pyannote.core.utils.distance import pdist
-from pyannote.core.utils.distance import l2_normalize
-
-from pyannote.audio.features import Precomputed
-from pyannote.audio.signal import Binarize, Peak
 
 from pyannote.core import Annotation
-from pyannote.core import SlidingWindowFeature
 from pyannote.database import get_annotated
 
-from pyannote.audio.embedding.clustering import HierarchicalPoolingClustering
 from pyannote.metrics.diarization import GreedyDiarizationErrorRate
 
-from .base import Pipeline
+from .speech_turn_segmentation import SpeechTurnSegmentation
+from .speech_turn_clustering import SpeechTurnClustering
+from .speech_turn_assignment import SpeechTurnClosestAssignment
+
+from typing import Optional
+from pyannote.pipeline import Pipeline
 import chocolate
 
 
-class NeuralSegmentation(Pipeline):
+class SpeakerDiarization(Pipeline):
+    """Speaker diarization pipeline
 
-    def __init__(self, sad=None, scd=None, **kwargs):
+    Hyper-parameters
+    ----------------
+    min_duration : `float`
+        Do not cluster speech turns shorter than `min_duration`. Assign them to
+        the closest cluster (of long speech turns) instead.
+    """
+
+    def __init__(self, sad_scores: Optional[Path] = None,
+                       scd_scores: Optional[Path] = None,
+                       embedding: Optional[Path] = None,
+                       metric: Optional[str] = 'cosine',
+                       method: Optional[str] = 'pool'):
+
         super().__init__()
-        self.sad = Path(sad).expanduser().resolve(strict=True)
-        self.scd = Path(scd).expanduser().resolve(strict=True)
-        self.with_params(**kwargs)
 
-    def get_tune_space(self):
-        return {
-            'sad_onset': chocolate.uniform(0., 1.),
-            'sad_offset': chocolate.uniform(0., 1.),
-            'scd_alpha': chocolate.uniform(0., 1.),
-            'scd_min_duration': chocolate.uniform(0., 5.),
-        }
+        self.sad_scores = sad_scores
+        self.scd_scores = scd_scores
+        self.speech_turn_segmentation = SpeechTurnSegmentation(
+            sad_scores=self.sad_scores,
+            scd_scores=self.scd_scores)
 
-    def get_tune_metric(self):
-        raise NotImplementedError()
+        self.min_duration = chocolate.uniform(0, 10)
 
-    def with_params(self, sad_onset=0.7, sad_offset=0.7,
-                    scd_alpha=0.5, scd_min_duration=1.):
+        self.embedding = embedding
+        self.metric = metric
+        self.method = method
+        self.speech_turn_clustering = SpeechTurnClustering(
+            embedding=self.embedding, metric=self.metric, method=self.method)
 
-        # initialize speech activity detection
-        self.sad_ = Precomputed(self.sad)
-        self.sad_onset = sad_onset
-        self.sad_offset = sad_offset
-        self.sad_binarize_ = Binarize(onset=sad_onset, offset=sad_offset)
+        self.speech_turn_assignment = SpeechTurnClosestAssignment(
+            embedding=self.embedding, metric=self.metric)
 
-        # initialize speaker change detection
-        self.scd_ = Precomputed(self.scd)
-        self.scd_alpha = scd_alpha
-        self.scd_min_duration = scd_min_duration
-        self.scd_peak_ = Peak(alpha=scd_alpha, min_duration=scd_min_duration)
+    def __call__(self, current_file: dict) -> Annotation:
+        """Apply speaker diarization
 
-        return self
+        Parameters
+        ----------
+        current_file : `dict`
+            File as provided by a pyannote.database protocol.
 
-    def apply(self, current_file):
+        Returns
+        -------
+        hypothesis : `pyannote.core.Annotation`
+            Speaker diarization output.
+        """
 
-        # Speech Activity Detection
+        # segmentation into speech turns
+        speech_turns = self.speech_turn_segmentation(current_file)
 
-        # get raw SAD scores
-        soft_sad = self.sad_(current_file)
+        # in case there is one speech turn or less, there is no need to apply
+        # any kind of clustering approach.
+        if len(speech_turns) < 2:
+            return speech_turns
 
-        # check once and for all whether SAD scores are log-scaled
-        if not hasattr(self, 'sad_log_scale_'):
-            if np.nanmean(soft_sad.data) < 0:
-                self.sad_log_scale_ = True
+        # split short/long speech turns. the idea is to first cluster long
+        # speech turns (i.e. those for which we can trust embeddings) and then
+        # assign each speech turn to the closest cluster.
+        long_speech_turns = speech_turns.empty()
+        shrt_speech_turns = speech_turns.empty()
+        for segment, track, label in speech_turns.itertracks(yield_label=True):
+            if segment.duration < self.min_duration:
+                shrt_speech_turns[segment, track] = label
             else:
-                self.sad_log_scale_ = False
+                long_speech_turns[segment, track] = label
 
-        # get SAD probability
-        prob_sad = np.exp(soft_sad.data) if self.sad_log_scale_ \
-                   else soft_sad.data
+        # in case there are no long speech turn to cluster, we return the
+        # original speech turns (= shrt_speech_turns)
+        if len(long_speech_turns) < 1:
+            return speech_turns
 
-        # support both non-speech/speech & non-speech/single/overlap
-        prob_sad = 1. - prob_sad[:, 0]
-        prob_sad = SlidingWindowFeature(prob_sad, soft_sad.sliding_window)
+        # first: cluster long speech turns
+        long_speech_turns = self.speech_turn_clustering(current_file,
+                                                        long_speech_turns)
 
-        # binarization
-        hard_sad = self.sad_binarize_.apply(prob_sad)
+        # then: assign short speech turns to clusters
+        long_speech_turns.rename_labels(generator='string', copy=False)
 
-        # Speaker Change Detection
+        if len(shrt_speech_turns) > 0:
+            shrt_speech_turns.rename_labels(generator='int', copy=False)
+            shrt_speech_turns = self.speech_turn_assignment(current_file,
+                                                            shrt_speech_turns,
+                                                            long_speech_turns)
+        # merge short/long speech turns
+        return long_speech_turns.update(shrt_speech_turns, copy=False)
 
-        # get raw SCD scores
-        soft_scd = self.scd_(current_file)
+        # TODO. add GMM-based resegmentation
 
-        # check once and for all whether SCD scores are log-scaled
-        if not hasattr(self, 'scd_log_scale_'):
-            if np.nanmean(soft_scd.data) < 0:
-                self.scd_log_scale_ = True
-            else:
-                self.scd_log_scale_ = False
+    def loss(self, current_file: dict, hypothesis: Annotation) -> float:
+        """Compute diarization error rate
 
-        # get SCD probability
-        prob_scd = np.exp(soft_scd.data) if self.scd_log_scale_ \
-                   else soft_scd.data
+        Parameters
+        ----------
+        current_file : `dict`
+            File as provided by a pyannote.database protocol.
+        hypothesis : `pyannote.core.Annotation`
+            Speaker diarization output.
 
-        # take the final dimension
-        # (in order to support both classification and regression scores)
-        prob_scd = prob_scd[:, -1]
-        prob_scd = SlidingWindowFeature(prob_scd, soft_scd.sliding_window)
+        Returns
+        -------
+        error : `float`
+            Diarization error rate
+        """
 
-        # peak detection
-        hard_scd = self.scd_peak_.apply(prob_scd)
-
-        speech_turns = hard_scd.crop(hard_sad)
-
-        # only process the annotated part
-        speech_turns = speech_turns.crop(get_annotated(current_file))
-
-        return speech_turns
+        metric = GreedyDiarizationErrorRate(collar=0.0, skip_overlap=False)
+        reference  = current_file['annotation']
+        uem = get_annotated(current_file)
+        return metric(reference, hypothesis, uem=uem)
 
 
-class Yin2018(NeuralSegmentation):
+class Yin2018(SpeakerDiarization):
     """Speaker diarization pipeline introduced in Yin et al., 2018
 
-        Ruiqing Yin, Hervé Bredin, and Claude Barras. "Neural speech turn
-        segmentation and affinity propagation for speaker diarization".
-        Interspeech 2018.
+    Ruiqing Yin, Hervé Bredin, and Claude Barras. "Neural speech turn
+    segmentation and affinity propagation for speaker diarization".
+    Interspeech 2018.
 
     Parameters
     ----------
-    sad : str
+    sad_scores : `Path`
         Path to precomputed speech activity detection scores.
-    scd : str
-        Path to precomputed speaker change detection scores.
-    emb : str
+    scd_scores : `Path`
+        Path to precomputed speaker change detection scores
+    embedding : `Path
         Path to precomputed embeddings.
     metric : {'euclidean', 'cosine', 'angular'}, optional
-        Metric used for comparing embeddings. Defaults to 'angular'.
+        Metric used for comparing embeddings. Defaults to 'cosine'.
     """
+    def __init__(self, sad_scores: Optional[Path] = None,
+                       scd_scores: Optional[Path] = None,
+                       embedding: Optional[Path] = None,
+                       metric: Optional[str] = 'cosine'):
 
-    def __init__(self, sad=None, scd=None, emb=None, metric='angular', **kwargs):
-        self.emb = Path(emb).expanduser().resolve(strict=True)
-        self.metric = metric
-        super().__init__(sad=sad, scd=scd)
-        self.with_params(**kwargs)
+        super().__init__(sad_scores=sad_scores,
+                         scd_scores=scd_scores,
+                         embedding=embedding,
+                         metric=metric,
+                         method='affinity_propagation')
 
-    def get_tune_space(self):
-
-        base_space = super().get_tune_space()
-        space = {
-            'cls_damping': chocolate.uniform(0.5, 1.),
-            # FIXME: be smarter about this parameter
-            'cls_preference': chocolate.uniform(-8.0, 0.0)}
-        base_space.update(space)
-        return base_space
-
-    def get_tune_metric(self):
-
-        def func(reference, hypothesis, uem=None):
-
-            # heuristic to avoid wasting time computing DER
-            # when the proposed solution is obviously wrong
-            r_labels = reference.crop(uem).labels()
-            h_labels = hypothesis.crop(uem).labels()
-            if len(h_labels) > 100 * len(r_labels):
-                return 1.
-
-            metric = GreedyDiarizationErrorRate()
-            return metric(reference, hypothesis, uem=uem)
-
-        return func
-
-    def with_params(self, sad_onset=0.7, sad_offset=0.7,
-                    scd_alpha=0.5, scd_min_duration=1.,
-                    cls_preference=-7.0,
-                    cls_damping=0.8):
-
-        # initialize speech activity detection and speaker change detection
-        super().with_params(sad_onset=sad_onset, sad_offset=sad_offset,
-                            scd_alpha=scd_alpha,
-                            scd_min_duration=scd_min_duration)
-
-        # initialize speech turn embedding
-        self.emb_ = Precomputed(self.emb)
-
-        # initialize clustering module
-        self.cls_damping = cls_damping
-
-        self.cls_preference = cls_preference
-        # NOTE cls_preference could be a multiplicative factor of a default
-        # affinity value (e.g. median affinity value)
-        self.cls_ = sklearn.cluster.AffinityPropagation(
-            damping=cls_damping, preference=cls_preference,
-            affinity='precomputed', max_iter=200, convergence_iter=15)
-
-        # sklearn documentation: Preferences for each point - points with
-        # larger values of preferences are more likely to be chosen as
-        # exemplars. The number of exemplars, ie of clusters, is influenced by
-        # the input preferences value. If the preferences are not passed as
-        # arguments, they will be set to the median of the input similarities.
-
-        # NOTE one could set the preference value of each speech turn
-        # according to their duration. longer speech turns are expected to
-        # have more accurate embeddings, therefore should be prefered for
-        # exemplars
-
-        return self
-
-    def apply(self, current_file):
-
-        # initial segmentation
-        speech_turns = super().apply(current_file)
-
-        # initialize the hypothesized annotation
-        hypothesis = Annotation(uri=current_file['uri'])
-        if len(speech_turns) < 1:
-            return hypothesis
-
-        # this only happens during pipeline training
-        if 'annotation' in current_file:
-            # number of speech turns in reference
-            reference = current_file['annotation']
-            n_turns_true = len(list(reference.itertracks()))
-
-            # number of speech turns in hypothesis
-            uem = get_annotated(current_file)
-            n_turns_pred = len(speech_turns.crop(uem))
-
-            # don't even bother trying to cluster those speech turns
-            # as there are too many of those...
-            if n_turns_pred > 20 * n_turns_true:
-                return None
-
-        # get raw (sliding window) embeddings
-        emb = self.emb_(current_file)
-
-        # get one embedding per speech turn
-        # FIXME don't l2_normalize for any metric
-        fX = l2_normalize(np.vstack(
-            [np.sum(emb.crop(t, mode='loose'), axis=0) for t in speech_turns]))
-
-        # apply clustering
-        try:
-            affinity = -squareform(pdist(fX, metric=self.metric))
-            clusters = self.cls_.fit_predict(affinity)
-        except MemoryError as e:
-            # cannot compute affinity propagation
-            return None
-
-        for speech_turn, cluster in zip(speech_turns, clusters):
-            # HACK find why fit_predict returns NaN sometimes and fix it.
-            cluster = -1 if np.isnan(cluster) else cluster
-            hypothesis[speech_turn] = cluster
-
-        return hypothesis
-
-
-class HierachicalEmbeddingPooling(NeuralSegmentation):
-    """
-
-    Parameters
-    ----------
-    sad : str
-        Path to precomputed speech activity detection scores.
-    scd : str
-        Path to precomputed speaker change detection scores.
-    emb : str
-        Path to precomputed embeddings.
-    metric : {'euclidean', 'cosine', 'angular'}, optional
-        Metric used for comparing embeddings. Defaults to 'angular'.
-    """
-
-    def __init__(self, sad=None, scd=None, emb=None, metric='angular', **kwargs):
-        self.emb = emb
-        self.metric = metric
-        super().__init__(sad=sad, scd=scd)
-        self.with_params(**kwargs)
-
-    def get_tune_space(self):
-
-        base_space = super().get_tune_space()
-        space = {'cls_threshold': chocolate.uniform(0, 2.)}
-        base_space.update(space)
-        return base_space
-
-    def get_tune_metric(self):
-
-        def func(reference, hypothesis, uem=None):
-
-            # heuristic to avoid wasting time computing DER
-            # when the proposed solution is obviously wrong
-            r_labels = reference.crop(uem).labels()
-            h_labels = hypothesis.crop(uem).labels()
-            if len(h_labels) > 100 * len(r_labels):
-                return 1.
-
-            metric = GreedyDiarizationErrorRate()
-            return metric(reference, hypothesis, uem=uem)
-
-        return func
-
-    def with_params(self, sad_onset=0.7, sad_offset=0.7,
-                    scd_alpha=0.5, scd_min_duration=1.,
-                    cls_threshold=0.8):
-
-        # initialize speech activity detection and speaker change detection
-        super().with_params(sad_onset=sad_onset, sad_offset=sad_offset,
-                            scd_alpha=scd_alpha,
-                            scd_min_duration=scd_min_duration)
-
-        # initialize speech turn embedding
-        self.emb_ = Precomputed(self.emb)
-
-        # initialize clustering module
-        self.cls_threshold = cls_threshold
-
-        self.cls_ = HierarchicalPoolingClustering(metric=self.metric)
-
-        return self
-
-    def apply(self, current_file):
-
-        # (inherited) initial segmentation
-        seg = super().apply(current_file)
-
-        # load raw embeddings
-        emb = self.emb_(current_file)
-
-        # hierarchical pooling clustering
-        return self.cls_.fit(seg, emb).apply(self.cls_threshold)
+        self.freeze({
+            'min_duration': 0.,
+            'speech_turn_segmentation': {
+                'speech_activity_detection': {'min_duration_on': 0.,
+                                              'min_duration_off': 0.,
+                                              'pad_onset': 0.,
+                                              'pad_offset': 0.}}})