+++ pyannote/audio/pipeline/speech_turn_segmentation.py
@@ -1,0 +1,133 @@
+#!/usr/bin/env python
+# encoding: utf-8
+
+# The MIT License (MIT)
+
+# Copyright (c) 2018 CNRS
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+# AUTHORS
+# Hervé BREDIN - http://herve.niderb.fr
+
+from typing import Optional
+from pathlib import Path
+
+from pyannote.core import Annotation
+from pyannote.pipeline import Pipeline
+from .speaker_change_detection import SpeakerChangeDetection
+from .speech_activity_detection import SpeechActivityDetection
+
+from pyannote.database import get_annotated
+from pyannote.metrics.diarization import DiarizationPurityCoverageFMeasure
+
+
+class OracleSpeechTurnSegmentation(Pipeline):
+    """Oracle segmentation"""
+
+    def __call__(self, current_file: dict) -> Annotation:
+        """Return groundtruth segmentation
+
+        Parameter
+        ---------
+        current_file : `dict`
+            Dictionary as provided by `pyannote.database`.
+
+        Returns
+        -------
+        hypothesis : `pyannote.core.Annotation`
+            Speech turns
+        """
+
+        return current_file['annotation'].relabel_tracks(generator='string')
+
+
+class SpeechTurnSegmentation(Pipeline):
+    """Combine speech activity and speaker change detections for segmentation
+
+    Parameters
+    ----------
+    sad_scores : `Path`
+        Path to precomputed speech activity detection scores.
+    scd_scores : `Path`
+        Path to precomputed speaker change detection scores
+    purity : `float`, optional
+        Target purity. Defaults to 0.95
+    """
+
+    def __init__(self, sad_scores: Optional[Path] = None,
+                       scd_scores: Optional[Path] = None,
+                       purity: Optional[float] = 0.95):
+        super().__init__()
+
+        self.purity = purity
+
+        self.sad_scores = sad_scores
+        self.speech_activity_detection = SpeechActivityDetection(scores=self.sad_scores)
+
+        self.scd_scores = scd_scores
+        self.speaker_change_detection = SpeakerChangeDetection(scores=self.scd_scores)
+
+    def __call__(self, current_file: dict) -> Annotation:
+        """Apply speech turn segmentation
+
+        Parameter
+        ---------
+        current_file : `dict`
+            Dictionary as provided by `pyannote.database`.
+
+        Returns
+        -------
+        hypothesis : `pyannote.core.Annotation`
+            Hypothesized speech turns
+        """
+
+        sad = self.speech_activity_detection(current_file).get_timeline()
+        scd = self.speaker_change_detection(current_file)
+        speech_turns = scd.crop(sad, mode='intersection')
+        speech_turns.modality = 'speaker'
+        return speech_turns
+
+    def loss(self, current_file: dict, hypothesis: Annotation) -> float:
+        """Compute (1 - coverage) at target purity
+
+        If purity < target, return 1 + (1 - purity)
+
+        Parameters
+        ----------
+        current_file : `dict`
+            File as provided by a pyannote.database protocol.
+        hypothesis : `pyannote.core.Annotation`
+            Speech turns.
+
+        Returns
+        -------
+        error : `float`
+            1. - cluster coverage.
+        """
+
+        metric = DiarizationPurityCoverageFMeasure()
+        reference  = current_file['annotation']
+        uem = get_annotated(current_file)
+        f_measure = metric(reference, hypothesis, uem=uem)
+        purity, coverage, _ = metric.compute_metrics()
+        if purity > self.purity:
+            return 1. - coverage
+        else:
+            return 1. + (1. - purity)