+++ pyannote/audio/pipeline/utils.py
@@ -26,6 +26,35 @@
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
 
-from .speech_activity_detection import SpeechActivityDetection
-from .speech_turn_segmentation import SpeechTurnSegmentation
-from .speaker_diarization import SpeakerDiarization
+
+from pyannote.core import Annotation
+
+
+def assert_string_labels(annotation: Annotation, name: str):
+    """Check that annotation only contains string labels
+
+    Parameters
+    ----------
+    annotation : `pyannote.core.Annotation`
+        Annotation.
+    name : `str`
+        Name of the annotation (used for user feedback in case of failure)
+    """
+    if any(not isinstance(label, str) for label in annotation.labels()):
+        msg = f'{name} must contain `str` labels only.'
+        raise ValueError(msg)
+
+
+def assert_int_labels(annotation: Annotation, name: str):
+    """Check that annotation only contains integer labels
+
+    Parameters
+    ----------
+    annotation : `pyannote.core.Annotation`
+        Annotation.
+    name : `str`
+        Name of the annotation (used for user feedback in case of failure)
+    """
+    if any(not isinstance(label, int) for label in annotation.labels()):
+        msg = f'{name} must contain `int` labels only.'
+        raise ValueError(msg)