+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Version 1.1 (WIP)
 
+  - BREAKING: switch to new pyannote.pipeline package
   - BREAKING: add unified FeatureExtraction base class
   - feat: add support for on-the-fly data augmentation
   - setup: switch to librosa 0.6